refactor: eliminate gosec warnings through int32 type alignment
  Root Cause Fix: - Configure sqlc to generate int32 for PostgreSQL
  INTEGER types - Update SQL queries with explicit ::integer casts -
  Remove unnecessary safeInt32() overflow checking function

  Type System Improvements: - Align internal types with database schema
  (int32) - Update function signatures: limit, offset, topK → int32 -
  Update config fields: MaxHistoryMessages, RAGTopK → int32 -
  Synchronize all interfaces: SessionStore, IndexerStore, SessionQuerier

  Safety Documentation: - Add #nosec comments for unavoidable API
  boundary conversions - Document why conversions are safe (validated
  ranges, bounded values) - Add validation for user input (parsedLimit
  range check 1-1000)

  Test Updates: - Fix all mock implementations to match new int32
  signatures - Update test fixtures and table-driven test data types -
  Fix environment variable test (KOOPA_ prefix)

+++ db/queries/sessions.sql
@@ -41,7 +41,7 @@ LIMIT sqlc.arg(result_limit)
 OFFSET sqlc.arg(result_offset);
 
 -- name: GetMaxSequenceNumber :one
-SELECT COALESCE(MAX(sequence_number), 0) AS max_seq
+SELECT COALESCE(MAX(sequence_number), 0)::integer AS max_seq
 FROM session_messages
 WHERE session_id = $1;
